+++ src/service.rs
@@ -20,35 +20,13 @@ pub mod keyvalue;
 use std::any::Any;
 use std::collections::HashMap;
 
-use anyhow::Result;
 use wasmtime::StoreLimits;
-use wasmtime::component::{InstancePre, Linker};
 use wasmtime_wasi::{IoView, ResourceTable, WasiCtx, WasiCtxBuilder, WasiView};
 
 // use wasmtime_wasi_http::WasiHttpView;
 use crate::trace::{Errout, Stdout};
 
 pub type Metadata = Box<dyn Any + Send>;
-
-/// Service represents a particular runtime service depended on by wasm
-/// components. For example, an HTTP server or a message broker.
-pub trait Service: Sync + Send {
-    type Ctx: IoView + WasiView;
-
-    /// Returns the wasi namespace the service supports. For example, `wasi:http`.
-    fn namespace(&self) -> &'static str;
-
-    /// Add the service to the wasm component linker.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the service encounters an error adding generated
-    /// bindings to the linker.
-    fn add_to_linker(&self, linker: &mut Linker<Self::Ctx>) -> Result<()>;
-
-    /// Start and run the runtime.
-    fn start(&self, pre: InstancePre<Self::Ctx>) -> impl Future<Output = Result<()>> + Send;
-}
 
 /// Ctx implements messaging host interfaces. In addition, it holds the
 /// host-defined state used by the wasm runtime [`Store`].
