+++ src/service/http.rs
@@ -39,7 +39,7 @@ pub fn new() -> Service {
     Service { addr }
 }
 
-impl crate::Service for Service {
+impl runtime::Service for Service {
     type Ctx = Ctx;
 
     fn namespace(&self) -> &'static str {
